Refactor: create useFetch hooks and file
