+++ frontend/src/page/Milestone.js
@@ -4,21 +4,14 @@ import MilestoneList from '@Component/MilestonePage/MilestoneList'
 import { getMilestones } from '@Api/milestone' // TODO: API 다른 거로 바꿔야 함
 import styled from 'styled-components'
 import Button from '@Component/common/Button'
+import { useFetch } from '@Util/hook'
 
 export const milestoneContext = createContext()
-
-/* TODO: 위치 변경 */
-const fetchData = async (requestFn, setFn) => {
-  const response = await requestFn()
-  setFn(response)
-}
 
 const MilestonePage = () => {
   const [milestones, setMilestones] = useState([])
 
-  useEffect(() => {
-    fetchData(getMilestones, setMilestones)
-  }, [])
+  useFetch(getMilestones, setMilestones)
 
   return (
     <milestoneContext.Provider value={{ milestones, setMilestones }}>
