+++ frontend/src/util/hook.js
@@ -1,0 +1,14 @@
+import { useEffect } from 'react'
+
+const useFetch = (requestFn, setFn) => {
+  useEffect(() => {
+    fetchData(requestFn, setFn)
+  }, [])
+}
+
+const fetchData = async (requestFn, setFn) => {
+  const response = await requestFn()
+  setFn(response)
+}
+
+export { useFetch }
